fix bugs, add async support

+++ bin/repolinter.js
@@ -1,6 +1,12 @@
 #!/usr/bin/env node
 // Copyright 2017 TODO Group. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
+const path = require('path')
+const repolinter = require('..')
+const rimraf = require('rimraf')
+const git = require('simple-git/promise')()
+const fs = require('fs')
+const os = require('os')
 
 require('yargs')
   .command('lint <directory>', 'run repolinter on the specified directory, outputting results to STDOUT.', yargs => {
@@ -33,29 +39,26 @@ require('yargs')
         default: false,
         type: 'boolean'
       })
-  }, (argv) => {
-    const path = require('path')
-    const repolinter = require('..')
-    
+  }, async (/** @type {any} */ argv) => {
+    let tmpDir = null
+    // temporarily clone a git repo to lint
     if (argv.git) {
-      const git = require('simple-git')()
-      const uuidv4 = require('uuid/v4')
-      const rimraf = require('rimraf')
-      const tmpDir = path.resolve(process.cwd(), 'tmp', uuidv4())
-
-      git.clone(argv.directory, tmpDir, (error) => {
-        if (!error) {
-          const output = repolinter.lint(tmpDir, argv.allowPaths, argv.dryRun, argv.ruleset)
-          console.log(repolinter.defaultFormatter.formatOutput(output, argv.dryRun))
-          process.exitCode = output.passed ? 0 : 1
-        }
+      tmpDir = await fs.promises.mkdtemp(path.join(os.tmpdir(), 'repolinter-'))
+      const result = await git.clone(argv.directory, tmpDir)
+      if (result) {
+        console.error(result)
+        process.exitCode = 1
         rimraf(tmpDir, function () {})
-      })
-    } else {
-      const output = repolinter.lint(path.resolve(process.cwd(), argv.directory), argv.allowPaths, argv.dryRun, argv.ruleset)
-      console.log(repolinter.defaultFormatter.formatOutput(output, argv.dryRun))
-      process.exitCode = output.passed ? 0 : 1
+        return
+      }
     }
+    // run the linter
+    const output = await repolinter.lint(tmpDir || path.resolve(process.cwd(), argv.directory), argv.allowPaths, argv.dryRun, argv.ruleset)
+    console.log(repolinter.defaultFormatter.formatOutput(output, argv.dryRun))
+    process.exitCode = output.passed ? 0 : 1
+    // delete the tmpdir if it exists
+    if (tmpDir)
+      rimraf(tmpDir, function () {})
   })
   .demandCommand()
   .help()
